Split serialization into some separate files
